change defaults in database when needed
If a config value changes with respect to the default value in the
database set the new default also in the database. This makes changed
defaults as fitting the needs of the daemons easier.

Also protect the measurement lists with a mutex against interfernce,
the need was triggered by the changes mentioned above.