move work from usart ISR to fNextLine()
this way the ISR is quicker as it adds the bytes to the ring buffer,
the cost is that we now every time search the ring buffer for a newline.
But because there are usually no characters in the buffer that is very quick.
Also we update the volatile variables after copying out the line, thus
saving time spent with the interrupts disabled.